Fixed UI elements to scale properly with screen size changes.
Issue ive noticed is that when put into the main game scene, player is unable to click on buttons. I believe the tile selector or something is disallowing ui interaction. Once we can figure out how to change that, we can have it for pressing alt will switch the cursor from tile select to ui select. Or implement a deadzone for ui areas.